use uint32_t for table size

+++ src/table.h
@@ -11,7 +11,7 @@ using namespace std;
 class Table {
   public:
   map<string, Field *> fields;
-  size_t size; // record count
+  uint32_t size; // record count
 
   void setField(Field *field) {
     fields[field->name] = field;
